Fortify 21.2.x

+++ prepare.sh
@@ -1,46 +1,53 @@
 #!/bin/bash
 
-function install { 
-    echo "          "
-    echo "Prepare started"
-    echo "          "
-    FORTIFY_MAJOR=21.1.0
-    SOURCE_AND_LIB_SCANNER_MAJOR=21.1.0
-    WEBINSPECT_MAJOR=21.1
+function PrepareInstall { 
+    echo "################################################################################"
+    echo "#          Prepare to started"
+    echo "################################################################################"
+    FORTIFY_MAJOR=21.2.0
+    SOURCE_AND_LIB_SCANNER_MAJOR=21.1.2
+    WEBINSPECT_MAJOR=21.2
     
-    FORTIFY_SSC_MAJOR=21.1.2
-    FORTIFY_SCANCENTRAL_MAJOR=21.1.0
-    FORTIFY_SCA_MAJOR=21.1.2
-    FORTIFY_SCA_PLUGINS_MAJOR=21.1.1
+    FORTIFY_SSC_MAJOR=21.2.2
+    FORTIFY_SCANCENTRAL_MAJOR=21.2.3
+    FORTIFY_SCA_MAJOR=21.2.3
+    FORTIFY_SCA_PLUGINS_MAJOR=21.2.3
     
     TOMCAT_MAJOR=9
-    TOMCAT_VERSION=9.0.54
+    TOMCAT_VERSION=9.0.58
 
-    FILE_FORTIFY_LICESE=Fortify_2021.license
-    FILE_JDK=jdk-11.0.11_linux-x64_bin.tar.gz
+    FILE_FORTIFY_LICESE=Fortify_2022.license
+    FILE_JDK=jdk-11.0.14_linux-x64_bin.tar.gz
 
     FILE_FORTIFY=Fortify_${FORTIFY_MAJOR}.zip
-    FILE_SCANCENTRAL=Fortify_ScanCentral_Controller_${FORTIFY_MAJOR}.zip
-    FILE_SCANCENTRAL_CONTROLLER=Fortify_ScanCentral_Controller_${FORTIFY_MAJOR}_x64.zip
-    FILE_SCANCENTRAL_CLIENT=Fortify_ScanCentral_Client_${FORTIFY_MAJOR}_x64.zip
+    FILE_SSC_SERVER=Fortify_SSC_Server_${FORTIFY_SSC_MAJOR}.zip
+    FILE_SSC_WAR=Fortify_${FORTIFY_SSC_MAJOR}_Server_WAR_Tomcat.zip
+    FILE_SCANCENTRAL=Fortify_ScanCentral_Controller_${FORTIFY_SCANCENTRAL_MAJOR}.zip
+    FILE_SCANCENTRAL_CONTROLLER=Fortify_ScanCentral_Controller_${FORTIFY_SCANCENTRAL_MAJOR}_x64.zip
+    FILE_SCANCENTRAL_CLIENT=Fortify_ScanCentral_Client_${FORTIFY_SCANCENTRAL_MAJOR}_x64.zip
     FILE_SCA_LINUX=Fortify_SCA_and_Apps_${FORTIFY_SCA_MAJOR}_linux_x64.run
-    FILE_SCA_OSX=Fortify_SCA_and_Apps_${FORTIFY_SCA_MAJOR}_osx_x64.app.zip
     FILE_SCA_WINDOWS=Fortify_SCA_and_Apps_${FORTIFY_SCA_MAJOR}_windows_x64.exe
+    FILE_SCA_MAC=Fortify_SCA_and_Apps_${FORTIFY_SCA_MAJOR}_Mac.tar.gz
+    FILE_SCA_OSX=Fortify_SCA_and_Apps_${FORTIFY_SCA_MAJOR}_osx_x64.app.zip
     FILE_LIB_SCANNER=Fortify_SourceAndLibScanner_${SOURCE_AND_LIB_SCANNER_MAJOR}_x64.zip
     FILE_WI=WebInspect_64_${WEBINSPECT_MAJOR}.zip
     FILE_SECURITY_TOOLKIT=SecurityToolkit_${WEBINSPECT_MAJOR}.zip
     FILE_RUNTIME_AGENT=WebInspectAgent_${WEBINSPECT_MAJOR}.zip  
 
-    FILE_SSC_PATCH=Fortify_21.1.2_Server_WAR_Tomcat.zip
-    FILE_SCANCENTRAL_PATCH=Fortify_ScanCentral_Controller_21.1.1.zip
-    FILE_SCA_PATCH_LINUX=Fortify_SCA_and_Apps_21.1.2_linux_x64.zip
-    FILE_SCA_PATCH_OSX=Fortify_SCA_and_Apps_21.1.2_osx_x64.app.zip
-    FILE_SCA_PATCH_WINDOWS=Fortify_SCA_and_Apps_21.1.2_windows_x64.zip
+    FILE_SSC_PATCH=Fortify_21.2.2_Server.zip
+    FILE_SCANCENTRAL_PATCH=Fortify_ScanCentral_Controller_21.2.3.zip
+    FILE_SCA_PATCH_LINUX=Fortify_SCA_and_Apps_21.2.3_Linux.tar.gz
+    FILE_SCA_PATCH_OSX=Fortify_SCA_and_Apps_21.2.3_Mac.tar.gz
+    FILE_SCA_PATCH_WINDOWS=Fortify_SCA_and_Apps_21.2.3_Windows.zip
 
     FILE_TOMCAT=apache-tomcat-${TOMCAT_VERSION}.tar.gz
 
     FILE_BIRT=birt-report-designer-all-in-one-4.7.0-20170622-win32.win32.x86_64.zip
     URL_BIRT=http://www.eclipse.org/downloads/download.php?file=/birt/downloads/drops/R-R1-4.7.0-201706222054/${FILE_BIRT}
+
+    echo "Listing files at FortifyInstallers" 
+    ls -alF FortifyInstallers 
+    echo "          "
 
     echo "*** Unziping ${FILE_FORTIFY} to FortifyInstallers/ " 
     unzip -qq FortifyInstallers/${FILE_FORTIFY} -d FortifyInstallers/
@@ -63,23 +70,39 @@ function install {
     wget https://dlcdn.apache.org/tomcat/tomcat-connectors/native/1.2.31/source/tomcat-native-1.2.31-src.tar.gz -P FortifyInstallers/ 
     mv FortifyInstallers/tomcat-native-1.2.31-src.tar.gz FortifyInstallers/tomcat-native-latest.tar.gz
 
-    mkdir -p 00-fortify-centos8/setup/FortifyInstallers
-    mv FortifyInstallers/${FILE_JDK} 00-fortify-centos8/setup/FortifyInstallers/jdk-11_latest_linux-x64_bin.tar.gz
+    mkdir -p 00-fortify-core/setup/FortifyInstallers
+    mv FortifyInstallers/${FILE_JDK} 00-fortify-core/setup/FortifyInstallers/jdk-11_latest_linux-x64_bin.tar.gz
 
     mkdir -p 02-fortify-scancentral-sast/setup/FortifyInstallers    
     if [ -f "FortifyInstallers/${FILE_SCANCENTRAL_PATCH}" ]; then
-        mv FortifyInstallers/${FILE_SCANCENTRAL_PATCH} 02-fortify-scancentral-sast/setup/FortifyInstallers/Fortify_ScanCentral_Controller_Latest_x64.zip
+        if [[ "${FILE_SCANCENTRAL_PATCH}" == *"x64"* ]]; then
+            mv FortifyInstallers/${FILE_SCANCENTRAL_PATCH} 02-fortify-scancentral-sast/setup/FortifyInstallers/Fortify_ScanCentral_Controller_Latest_x64.zip
+        else
+            mkdir -p work/scsast/
+            echo "*** Unziping FortifyInstallers/${FILE_SCANCENTRAL_PATCH} to work/scsast/ " 
+            unzip -qq FortifyInstallers/${FILE_SCANCENTRAL_PATCH} -d work/scsast/
+            mv work/scsast/${FILE_SCANCENTRAL_CONTROLLER} 02-fortify-scancentral-sast/setup/FortifyInstallers/Fortify_ScanCentral_Controller_Latest_x64.zip
+            mv work/scsast/${FILE_SCANCENTRAL_CLIENT} FortifyInstallers/Fortify_ScanCentral_Client_Latest_x64.zip
+        fi
     else
         mv FortifyInstallers/${FILE_SCANCENTRAL_CONTROLLER} 02-fortify-scancentral-sast/setup/FortifyInstallers/Fortify_ScanCentral_Controller_Latest_x64.zip
     fi
-    cp FortifyInstallers/tomcat-native-latest.tar.gz 02-fortify-scancentral-sast/setup/FortifyInstallers/
 
-    mv FortifyInstallers/${FILE_SCANCENTRAL_CLIENT} FortifyInstallers/Fortify_ScanCentral_Client_Latest_x64.zip
+    if [ -f "FortifyInstallers/${FILE_SCANCENTRAL_CLIENT}" ]; then
+        mv FortifyInstallers/${FILE_SCANCENTRAL_CLIENT} FortifyInstallers/Fortify_ScanCentral_Client_Latest_x64.zip
+    fi
+    
+    cp FortifyInstallers/tomcat-native-latest.tar.gz 02-fortify-scancentral-sast/setup/FortifyInstallers/
 
     mkdir -p work/sca-linux/
     mkdir -p 03-fortify-sca-linux/setup/FortifyInstallers
     if [ -f "FortifyInstallers/${FILE_SCA_PATCH_LINUX}" ]; then
-        unzip -qq FortifyInstallers/${FILE_SCA_PATCH_LINUX} -d work/sca-linux/
+         echo "*** Unziping ${FILE_SCA_PATCH_LINUX} to work/sca-linux/ " 
+        if [[ "FortifyInstallers/${FILE_SCA_PATCH_LINUX}" =~ \.zip$ ]]; then            
+            unzip -qq FortifyInstallers/${FILE_SCA_PATCH_LINUX} -d work/sca-linux/
+        else
+            tar -pxzf FortifyInstallers/${FILE_SCA_PATCH_LINUX} -C work/sca-linux/ 
+        fi
     else
         tar -pxzf FortifyInstallers/Fortify_SCA_and_Apps_${FORTIFY_SCA_MAJOR}_Linux.tar.gz -C work/sca-linux/ 
     fi
@@ -89,29 +112,51 @@ function install {
     mkdir -p work/sca-windows/
     mkdir -p 04-fortify-sca-windows/setup/FortifyInstallers
     if [ -f "FortifyInstallers/${FILE_SCA_PATCH_WINDOWS}" ]; then
+        echo "*** Unziping ${FILE_SCA_PATCH_WINDOWS} to work/sca-windows/ " 
         unzip -qq FortifyInstallers/${FILE_SCA_PATCH_WINDOWS} -d work/sca-windows/
     else
+        echo "*** Unziping Fortify_SCA_and_Apps_${FORTIFY_SCA_MAJOR}_Windows.zip to work/sca-windows/ " 
         unzip -qq FortifyInstallers/Fortify_SCA_and_Apps_${FORTIFY_SCA_MAJOR}_Windows.zip -d work/sca-windows/
     fi
     mv work/sca-windows/${FILE_SCA_WINDOWS} FortifyInstallers/Fortify_SCA_Windows_Latest_x64.exe
     cp FortifyInstallers/Fortify_SCA_Windows_Latest_x64.exe 04-fortify-sca-windows/setup/FortifyInstallers/
 
-    if [ -f "FortifyInstallers/${FILE_SCA_PATCH_OSX}" ]; then
-        mv FortifyInstallers/${FILE_SCA_PATCH_OSX} FortifyInstallers/Fortify_SCA_OSX_Latest_x64.app.zip
+    mkdir -p work/sca-osx/
+    if [ -f "FortifyInstallers/${FILE_SCA_PATCH_OSX}" ]; then         
+        if [[ "FortifyInstallers/${FILE_SCA_PATCH_OSX}" =~ \.gz$ ]]; then 
+            tar -pxzf FortifyInstallers/${FILE_SCA_PATCH_OSX} -C work/sca-osx/
+        else
+            echo "*** Renaming ${FILE_SCA_PATCH_OSX} to Fortify_SCA_OSX_Latest_x64.app.zip " 
+            mv FortifyInstallers/${FILE_SCA_PATCH_OSX} FortifyInstallers/Fortify_SCA_OSX_Latest_x64.app.zip
+        fi
     else
-        mkdir -p work/sca-osx/
-        tar -pxzf FortifyInstallers/Fortify_SCA_and_Apps_${FORTIFY_SCA_MAJOR}_Mac.tar.gz -C work/sca-osx/
+        tar -pxzf FortifyInstallers/${FILE_SCA_MAC} -C work/sca-osx/        
+    fi
+
+    if [ -f "work/sca-osx/${FILE_SCA_OSX}" ]; then
         mv work/sca-osx/${FILE_SCA_OSX} FortifyInstallers/Fortify_SCA_OSX_Latest_x64.app.zip
     fi
 
     mv FortifyInstallers/${FILE_LIB_SCANNER} FortifyInstallers/SourceAndLibScanner_Latest_x64.zip
     
-    mv FortifyInstallers/Fortify_IntelliJ_Remediation_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip FortifyInstallers/Fortify_IntelliJ_Remediation_Plugin_Latest.zip
+    if [ -f "FortifyInstallers/Fortify_IntelliJ_Remediation_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip" ]; then 
+        mv FortifyInstallers/Fortify_IntelliJ_Remediation_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip FortifyInstallers/Fortify_IntelliJ_Remediation_Plugin_Latest.zip
+    elif [ -f "work/sca-windows/Fortify_IntelliJ_Remediation_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip" ]; then 
+        mv work/sca-windows/Fortify_IntelliJ_Remediation_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip FortifyInstallers/Fortify_IntelliJ_Remediation_Plugin_Latest.zip
+    fi
 
-    mv FortifyInstallers/Fortify_Eclipse_Remediation_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip FortifyInstallers/Fortify_Eclipse_Remediation_Plugin_Latest.zip
+    if [ -f "FortifyInstallers/Fortify_Eclipse_Remediation_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip" ]; then 
+        mv FortifyInstallers/Fortify_Eclipse_Remediation_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip FortifyInstallers/Fortify_Eclipse_Remediation_Plugin_Latest.zip
+    elif [ -f "work/sca-windows/Fortify_Eclipse_Remediation_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip" ]; then 
+        mv work/sca-windows/Fortify_Eclipse_Remediation_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip FortifyInstallers/Fortify_Eclipse_Remediation_Plugin_Latest.zip
+    fi
 
-    mv FortifyInstallers/Fortify_SecurityAssistant_Eclipse_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip FortifyInstallers/Fortify_SecurityAssistant_Eclipse_Plugin_Latest.zip
-
+    if [ -f "FortifyInstallers/Fortify_SecurityAssistant_Eclipse_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip" ]; then 
+        mv FortifyInstallers/Fortify_SecurityAssistant_Eclipse_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip FortifyInstallers/Fortify_SecurityAssistant_Eclipse_Plugin_Latest.zip
+    elif [ -f "work/sca-windows/Fortify_SecurityAssistant_Eclipse_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip" ]; then 
+        mv work/sca-windows/Fortify_SecurityAssistant_Eclipse_Plugin_${FORTIFY_SCA_PLUGINS_MAJOR}.zip FortifyInstallers/Fortify_SecurityAssistant_Eclipse_Plugin_Latest.zip
+    fi
+    
     mv FortifyInstallers/${FILE_WI} FortifyInstallers/WebInspect_64_Latest.zip
 
     mv FortifyInstallers/${FILE_SECURITY_TOOLKIT} FortifyInstallers/SecurityToolkit_Latest.zip
@@ -119,11 +164,21 @@ function install {
     mv FortifyInstallers/${FILE_RUNTIME_AGENT} FortifyInstallers/WebInspectAgent_Latest.zip  
 
     if [ -f "FortifyInstallers/${FILE_SSC_PATCH}" ]; then
-        mv FortifyInstallers/${FILE_SSC_PATCH} FortifyInstallers/Fortify_Latest_Server_WAR_Tomcat.zip
+        if [[ "${FILE_SSC_PATCH}" == *"WAR"* ]]; then
+            mv FortifyInstallers/${FILE_SSC_PATCH} FortifyInstallers/Fortify_Latest_Server_WAR_Tomcat.zip
+        else
+            mkdir -p work/ssc/
+            echo "*** Unziping FortifyInstallers/${FILE_SSC_PATCH} to work/ssc/ " 
+            unzip -qq FortifyInstallers/${FILE_SSC_PATCH} -d work/ssc/
+        fi
     else
         mkdir -p work/ssc/
-        unzip -qq FortifyInstallers/Fortify_SSC_Server_${FORTIFY_SCA_MAJOR}.zip -d work/ssc/
-        mv work/ssc/Fortify_${FORTIFY_SSC_MAJOR}_Server_WAR_Tomcat.zip FortifyInstallers/Fortify_Latest_Server_WAR_Tomcat.zip
+        echo "*** Unziping FortifyInstallers/${FILE_SSC_SERVER} to work/ssc/ " 
+        unzip -qq FortifyInstallers/${FILE_SSC_SERVER} -d work/ssc/        
+    fi
+
+    if [ -f "work/ssc/${FILE_SSC_WAR}" ]; then
+        mv work/ssc/${FILE_SSC_WAR} FortifyInstallers/Fortify_Latest_Server_WAR_Tomcat.zip
     fi
 
     cp -rf FortifyInstallers 01-fortify-ssc/setup/
@@ -148,9 +203,13 @@ function install {
         fi
     fi
 
+    echo "Updated list of files at FortifyInstallers" 
+    ls -alF FortifyInstallers 
     echo "          "
-    echo "Prepare finished"
-    echo "          "
+
+    echo "################################################################################"
+    echo "#          Preparation finished "
+    echo "################################################################################"
 }
 
-    install > ./prepare-install.log
+    PrepareInstall > ./prepare-install.log
